server/requirements.txt
Django
psycopg2-binary>=2.8
djangorestframework
djangorestframework-simplejwt
djoser